refactor: Coupon 컴포넌트 내부에서만 사용되는 유틸 함수를 Coupon 디렉터리로 이동
